requirements.txt
python-dotenv
mmpy_bot
schedule
requests
DateTime
python-time